edit 0-print_listint.c, 1-listint.c

+++ 0x13-more_linked_lists/0-print_listint.c
@@ -12,7 +12,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	size_t count;
+	int count;
 	const listint_t *ptr = h;
 
 	count = 0;
